# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sooqa API
  version: 0.1.0
  description: Private single-admin HTTP API for the durable sooqa media pipeline.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/v1/media:
    get:
      tags: [media]
      operationId: searchMedia
      summary: Search normalized media
      security:
        - bearerAuth: []
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: tags, in: query, description: Comma-separated normalized tags., schema: { type: string } }
        - { name: kind, in: query, schema: { $ref: '#/components/schemas/MediaKind' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/MediaStatus', default: active } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: A page of media rows.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/media/{id}:
    parameters:
      - $ref: '#/components/parameters/MediaId'
    get:
      tags: [media]
      operationId: getMedia
      summary: Read one normalized media row
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Media row.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Media' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      tags: [media]
      operationId: updateMedia
      summary: Update editorial media fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaUpdateRequest' }
      responses:
        '200':
          description: Updated media row.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Media' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/media/{id}/archive:
    post:
      tags: [media]
      operationId: archiveMedia
      summary: Archive media
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MediaId'
      responses:
        '200':
          description: Archived media row.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Media' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/media/{id}/tags:
    post:
      tags: [media]
      operationId: addMediaTag
      summary: Add a normalized tag
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MediaId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagRequest' }
      responses:
        '200':
          description: The attached tag.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/media/{id}/tags/{tag}:
    delete:
      tags: [media]
      operationId: removeMediaTag
      summary: Remove a normalized tag
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MediaId'
        - { name: tag, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Tag removed. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sooqa-api-token
  parameters:
...
    MediaId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Idempotency or state conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    PayloadTooLarge:
      description: Request body exceeds the configured limit.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Unexpected server failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, details]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
...
    MediaKind:
      type: string
      enum: [video, image, animation, audio]
    MediaStatus:
      type: string
      enum: [active, archived, deleted]
    Media:
      type: object
      required: [id, kind, status, tags, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/MediaKind' }
        status: { $ref: '#/components/schemas/MediaStatus' }
        title: { type: string, nullable: true }
        description: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        storage_state: { type: string, enum: [pending_storage, ready, storage_unknown, missing], nullable: true }
        source_url: { type: string, nullable: true }
        source_metadata: { type: object, additionalProperties: true, nullable: true }
        mime_type: { type: string, nullable: true }
        container: { type: string, nullable: true }
        video_codec: { type: string, nullable: true }
        audio_codec: { type: string, nullable: true }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        duration_ms: { type: integer, format: int64, nullable: true }
        bit_rate: { type: integer, format: int64, nullable: true }
        file_size_bytes: { type: integer, format: int64, nullable: true }
        sha256: { type: string, nullable: true }
        tags: { type: array, items: { $ref: '#/components/schemas/Tag' } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        archived_at: { type: string, format: date-time, nullable: true }
    MediaPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Media' } }
        next_cursor: { type: string, nullable: true }
    MediaUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        preferred_title: { type: string, nullable: true }
        editorial_description: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        expected_updated_at: { type: string, format: date-time, nullable: true }
    TagRequest:
      type: object
      required: [tag]
      additionalProperties: false
      properties:
        tag: { type: string, minLength: 1 }
    Tag:
      type: object
      required: [id, normalized_name, display_name, created_at]
      properties:
        id: { type: string, format: uuid }
        normalized_name: { type: string }
        display_name: { type: string }
        created_at: { type: string, format: date-time }
...